test: add tests for new pages flow
